+++ src/libtiled/tmxmapformat.h
@@ -2,27 +2,35 @@
  * tmxmapformat.h
  * Copyright 2008-2015, Thorbjørn Lindeijer <bjorn@lindeijer.nl>
  *
- * This file is part of Tiled.
+ * This file is part of libtiled.
  *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation; either version 2 of the License, or (at your option)
- * any later version.
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
  *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
+ *    1. Redistributions of source code must retain the above copyright notice,
+ *       this list of conditions and the following disclaimer.
  *
- * You should have received a copy of the GNU General Public License along with
- * this program. If not, see <http://www.gnu.org/licenses/>.
+ *    2. Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE CONTRIBUTORS ``AS IS'' AND ANY EXPRESS OR
+ * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
+ * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
+ * EVENT SHALL THE CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+ * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
 #pragma once
 
 #include "mapformat.h"
 #include "objecttemplateformat.h"
-#include "tilededitor_global.h"
+#include "tiled_global.h"
 #include "tilesetformat.h"
 
 namespace Tiled {
@@ -33,7 +41,7 @@ class MapObject;
 /**
  * A reader and writer for Tiled's .tmx map format.
  */
-class TILED_EDITOR_EXPORT TmxMapFormat : public MapFormat
+class TILEDSHARED_EXPORT TmxMapFormat : public MapFormat
 {
     Q_OBJECT
     Q_INTERFACES(Tiled::MapFormat)
@@ -77,7 +85,7 @@ private:
 /**
  * A reader and writer for Tiled's .tsx tileset format.
  */
-class TILED_EDITOR_EXPORT TsxTilesetFormat : public TilesetFormat
+class TILEDSHARED_EXPORT TsxTilesetFormat : public TilesetFormat
 {
     Q_OBJECT
     Q_INTERFACES(Tiled::TilesetFormat)
@@ -104,7 +112,7 @@ private:
 /**
  * A reader and writer for Tiled's .tx template format.
  */
-class TILED_EDITOR_EXPORT XmlObjectTemplateFormat : public ObjectTemplateFormat
+class TILEDSHARED_EXPORT XmlObjectTemplateFormat : public ObjectTemplateFormat
 {
     Q_OBJECT
     Q_INTERFACES(Tiled::ObjectTemplateFormat)